Exercicios1_24/Exercise_14/1.app_funtion_azurite_Insert_QUEUEs_BD/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
requests   
azure-storage-queue 
azure-storage-blob
uuid
python-dotenv
pymssql


